More fixes since I didn't look at all the files after directory restructure.
