+++ src/features/pokedata/redux/routeSlice.ts
@@ -1,4 +1,4 @@
-import PokeApi from '@/src/app/shared/api/PokeApiClient';
+import PokeApi from '@/src/shared/api/PokeApiClient';
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
