+++ src/shared/api/PokeApiClient.ts
@@ -1,4 +1,4 @@
-import { Berry, LocationArea } from '@/src/app/constants/types';
+import { Berry, LocationArea } from '@/src/constants/types';
 
 class PokeApiClient {
     private baseUrl = "https://pokeapi.co/api/v2";
